+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-06-10 22:36:29
-// Commit: feat(perf): add async perf stream provider
+// Updated at: 2026-06-11 12:11:47
+// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
 
 class perfComponentState {
-  final String commitRef = '1886';
-  final String lastMessage = 'feat(perf): add async perf stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-10T22:36:29+05:00');
+  final String commitRef = '1897';
+  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
+  final DateTime timestamp = DateTime.parse('2026-06-11T12:11:47+05:00');
 }